Commented out emitting errors in location update stream when location is not available
This was added by the MarsXan fork in 2018 (we never used that fork) but the original library would never notify on this stream when location was not available as the orignial APIs didn’t expose the `onLocationAvailability` callback.

This change makes this stream behave as the original implementation. In the future, if we’d like to expose location availability we can change the observable emission type or create a separeate stream to track availability exclusively.